Finished implementing saving effects
Now it is able to save effects for the larger area effects. Which may impact multiple specific people or objects.

This also involved putting the core of what is used to save the effects into its own function, so that it could be called as many times as necessary for an effect.